pyproject.toml
[tool.poetry]
name = "dama-bot"
version = "0.1.0"
description = ""
authors = ["Van1s1mys <malkovsin@gmail.com>"]
readme = "README.md"
packages = [
    {include = "db"},
    {include = "handlers"},
    {include = "bot"},
    {include = "services"}
]

[tool.poetry.dependencies]
python = ">=3.12"
sqlalchemy = "^2.0.40"
pandas = "^2.2.3"
psycopg2-binary = "^2.9.10"
openpyxl = "^3.1.5"
python-dotenv = "^1.1.0"
asyncpg = "^0.30.0"
aiogram = "^3.19.0"
redis = "^6.0.0"
openai = "^1.72.0"
aiohttp = "^3.11.16"
minio = "^7.2.15"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"